Fix issue of getting wrong asset id for watch
The ordering of asset was wrong causing issues in fetching the data.
Added batchTime as a group ordering so that the latest ingested data comes up first.
Refactored the retention time to be only on the config and rest from shell with same value.
Changed config to prod
Added extra error check of read timeout
Extended schedule slightly
Fixed gameId to be the actual gameCode
The UI ultimately will decide on which schedule to select based on availability
Refactor code to include constants for index name